AmazonCallback test refactoring
